Move base classes to apollo-common
