+++ packages/apollo-collaboration-server/src/utils/parse-change.pipe.ts
@@ -4,7 +4,7 @@ import {
   Injectable,
   PipeTransform,
 } from '@nestjs/common'
-import { Change, SerializedChange } from 'apollo-shared'
+import { Change, SerializedChange } from 'apollo-common'
 
 @Injectable()
 export class ParseChangePipe
