+++ packages/apollo-common/package.json
@@ -11,9 +11,30 @@
     "build": "tsc"
   },
   "dependencies": {
-    "@jbrowse/core": "^2.3.4"
+    "@gmod/gff": "^1.3.0",
+    "@jbrowse/core": "^2.3.4",
+    "apollo-schemas": "workspace:^",
+    "bson-objectid": "^2.0.4",
+    "tslib": "^2.3.1"
+  },
+  "peerDependencies": {
+    "@mui/material": "^5.0.0",
+    "@mui/x-data-grid": "^5.0.0",
+    "mobx": "^6.0.0",
+    "mobx-react": "^7.0.0",
+    "mobx-state-tree": "^5.0.0",
+    "prop-types": "^15.0.0",
+    "react": ">=16.8.0",
+    "react-dom": ">=16.8.0",
+    "rxjs": "^7.0.0",
+    "tss-react": "^4.0.0"
   },
   "devDependencies": {
+    "@nestjs/common": "^9.3.9",
+    "@nestjs/core": "^9.3.9",
+    "@types/node": "^18.14.2",
+    "apollo-mst": "workspace:^",
+    "mongoose": "^7.0.0",
     "typescript": "^4.7.4"
   },
   "packageManager": "yarn@3.2.2.",
