+++ packages/apollo-common/src/FeatureChange.ts
