+++ packages/apollo-common/src/Operation.ts
