+++ packages/apollo-common/src/Validation.ts
@@ -1,8 +1,9 @@
 import type { ExecutionContext } from '@nestjs/common'
 import type { Reflector } from '@nestjs/core'
-import { Change, ClientDataStore } from 'apollo-common'
 import { FeatureDocument } from 'apollo-schemas'
 import { ClientSession, Model } from 'mongoose'
+
+import { Change, ClientDataStore } from './Change'
 
 export interface Context {
   context: ExecutionContext
