+++ packages/apollo-common/src/index.ts
@@ -1,1 +1,7 @@
-export {}
+export * from './AssemblySpecificChange'
+export * from './Change'
+export * from './ChangeTypeRegistry'
+export * from './FeatureChange'
+export * from './Operation'
+export * from './OperationTypeRegistry'
+export * from './Validation'
