--- packages/apollo-shared/src/Changes/abstract/index.ts
@@ -1,3 +1,0 @@
-export * from './AssemblySpecificChange'
-export * from './Change'
-export * from './FeatureChange'
