+++ packages/apollo-shared/src/Changes/index.ts
@@ -1,4 +1,3 @@
-import { Change } from './abstract'
 import { AddAssemblyAndFeaturesFromFileChange } from './AddAssemblyAndFeaturesFromFileChange'
 import { AddAssemblyFromFileChange } from './AddAssemblyFromFileChange'
 import { AddFeatureChange } from './AddFeatureChange'
@@ -42,28 +41,3 @@ export * from './LocationEndChange'
 export * from './LocationStartChange'
 export * from './TypeChange'
 export * from './UserChange'
-
-// eslint-disable-next-line @typescript-eslint/no-explicit-any
-type ChangeType = new (...args: any[]) => Change
-
-class ChangeTypeRegistry {
-  changes: Map<string, ChangeType> = new Map()
-
-  registerChange(name: string, changeType: ChangeType): void {
-    if (this.changes.has(name)) {
-      throw new Error(`change type "${name}" has already been registered`)
-    }
-    this.changes.set(name, changeType)
-  }
-
-  getChangeType(name: string): ChangeType {
-    const RegisteredChangeType = this.changes.get(name)
-    if (!RegisteredChangeType) {
-      throw new Error(`No change constructor registered for "${name}"`)
-    }
-    return RegisteredChangeType
-  }
-}
-
-/** global singleton of all known types of changes */
-export const changeRegistry = new ChangeTypeRegistry()
