--- packages/apollo-shared/src/Operations/abstract/index.ts
@@ -1,1 +1,0 @@
-export * from './Operation'
