+++ packages/apollo-shared/src/index.ts
@@ -1,6 +1,4 @@
 export * from './Changes'
-export * from './Changes/abstract'
 export * from './Operations'
-export * from './Operations/abstract'
 export * from './Validations'
 export * from './Common'
